requirements.txt
aiogram>=3.4.1
asyncpg>=0.29.0
google-generativeai>=0.4.1
python-dotenv>=1.0.1
pydantic>=2.7.0
